# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ParkIt API",
        "description": "API for managing parkings and authentication in the ParkIt application",
        "contact": {
            "name": "API Support",
            "email": "support@parkit.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com"
        }
    ],
    "paths": {
...
        "/api/places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Get all places",
                "operationId": "98b7b5e708ea1544e70cf25ec8621387",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "parking_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "number": {
                                                "type": "string",
                                                "example": "A1"
                                            },
                                            "hourly_price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5.5,
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Store a new place",
                "operationId": "9b1406d5892a139db93bd35d7a248450",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parking_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "A1"
                                    },
                                    "hourly_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Place created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "parking_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "number": {
                                            "type": "string",
                                            "example": "A1"
                                        },
                                        "hourly_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5.5,
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/places/{id}": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Get a specific place",
                "operationId": "4ee90b2a793be42613fcf8ae54ff0cd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "parking_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "number": {
                                            "type": "string",
                                            "example": "A1"
                                        },
                                        "hourly_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5.5,
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Places"
                ],
                "summary": "Update a place",
                "operationId": "3eba858ed63a31cf3380c85c3bcd4ea2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parking_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "A1"
                                    },
                                    "hourly_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Place updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "parking_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "number": {
                                            "type": "string",
                                            "example": "A1"
                                        },
                                        "hourly_price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5.5,
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Places"
                ],
                "summary": "Delete a place",
                "operationId": "0b4281911812317e70e0ed38b948fa6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Place deleted successfully"
                    }
                }
            }
        },
        "/api/places/search/{searchKey}": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Search for places by number",
                "operationId": "e56b3e8ad74364f87406dc6c76dfe296",
                "parameters": [
                    {
                        "name": "searchKey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "parking_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "number": {
                                                "type": "string",
                                                "example": "A1"
                                            },
                                            "hourly_price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5.5,
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}